+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - 導入 GitHub Actions 作為 CI/CD 工具。([#b4b6b8c][b4b6b8c])
 - 建置本地 PlayWright 容器和 GitHub Actions 測試腳本。([#5502e70][5502e70])
 - 使用容器 web-deploy 架設 Nginx 伺服器。([#e034ad1][e034ad1])
+- 修正沒有安裝 postcss-import、postcss-nesting，導致 Playwright 測試失敗的問題。
 
 [e034ad1]: https://github.com/sugky7302/my-chat/commit/e034ad1f125f9143aef042aac5e7621f8499bc35
 [5502e70]: https://github.com/sugky7302/my-chat/commit/5502e70e0454a52483383c081109d8457cb734c2